LayoutTests/http/tests/security/contentSecurityPolicy/script-src-none-inline-event-expected.txt

CONSOLE MESSAGE: line 3: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS
